+++ CHANGELOG.md
@@ -29,6 +29,17 @@ Release Template
 # Releases
 Test/interim releases are not documented here.
 
+---
+
+## [0.17.1]
+### Fixed (minor)
+- Fix draw.io formatted (html) label conversion to text for new lines.
+  - https://github.com/StateSmith/StateSmith/issues/414
+- Fix multiple line guard code splitting across multiple lines.
+  - https://github.com/StateSmith/StateSmith/issues/415
+
+---
+
 ## [0.17.0]
 ### Added
 - Add support for idiomatic `Cpp` transpiler (only supports algorithm `Balanced2` right now).